requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
anthropic==0.16.0
stripe==7.11.0
requests==2.31.0
beautifulsoup4==4.12.2
SQLAlchemy==2.0.23
aiohttp==3.9.1
PyPDF2==3.0.1
python-magic==0.4.27
pydantic==2.5.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-digitalocean==1.17.0
dnspython==2.4.2
boto3==1.35.90
psycopg2-binary==2.9.9
